Add i18n for the monsters in the creatures sample mod
This has to be done in the mod side, instead of directly by EasyForger.
Technically, this is the case because we need the modId to locate the
i18n file, which only exists for the user created mod.

The con for this solution is the the user has to add the i18n file, if she
wants to avoid the un-i18n string.

The pro is that it is very easy to the user to customize the monster name.

closes #74